requirements.txt
google-cloud-bigquery>=3.33.0
